requirements.txt
numpy>=1.24
pandas>=2.0
tqdm>=4.66
Pillow>=10.0
matplotlib>=3.7
scikit-learn>=1.3
joblib>=1.3
torch>=2.1
torchvision>=0.16
